Add Ansible restart notifications. (#5)
When the Kuma binaries are updated, send a notification that can be
handled to restart any dependent services. Since Ansible notifications
are scoped to a play, this forces us to consolidate the binary installation
and service configuration roles to run in the same play. If we install
Kuma binaries for the "all" host group, they won't be modified for the
"control-plane" group, and the notification will never trigger.

Signed-off-by: James Peach <james.peach@konghq.com>