Make a first pass at a dataplane role.
The "data-plane" role is intended to configure a node as a Kuma dataplane
running a single service. This is more restrictive that we really want,
which is to be able to run services anywhere, so maybe at some point
this role should simply be included in the corresponding service role.

Dataplanes are not yet able to boostrap since they don't have a way to
authenticate to the control plane to get a dataplane token.

Added an "all" role to configure a generic control-plane DNS name for
all the nodes. This makes the kuma-cp instance reachable.

Various ports and such are still hard-coded across roles, and TLS is
still not enabled for the kuma-cp API server.

Signed-off-by: James Peach <james.peach@konghq.com>